// gameLogd/src/app/models/app.model.ts
export interface App {
    id: string;
    title: string;
    description: string;
    platforms: string[];
    releaseDate: string;
    developer: string;
    publisher: string;
    downloads: number;
    rating: number;
    imageUrl: string;
    totalRatingScore: number;
    numRatings: number;
    category: string;
    price: number;
    size: string;
    version: string;
} 